Reinstalled RESTFUL Authentication plugin and reupdated the models. Oh and I FINALLY GOT THE AUTHENTICATION TOKEN WORKING!
Things that changed in the configuration: 
We are no longer launching the application at http://localhost:3000/bin/Matlabulchifai.html
It is now changed to http://localhost:3000/matlabulchifai
Be sure to update your flex run and debug paths to the new url.

I created a new controller that launches the flex application through a ruby view.  This is the only way I found that reliably passes the authentication token into the flex app.  I'm also guessing it also allows us to force any requests to load a page from rails into the flex view... I'm not sure how to do this, so maybe Tommy can take a look at it.

My RadRails IDE shows one syntax error in index.html.erb for the flex view.  I have no idea why it it shows up or if will even be a problem for you guys.  If it does give you grief let me know and I'll try taking another look at it.  As it stands now the thing runs fine and it doesn't give any browser errors when the page loads... go figure.

I went through and updated as many POST url requests as I could with the authentication token.  If I missed any you'll get some error in flex after a post request.  I know there are still some things that need to be updated in patient summary, so there might be some phantom errors there.

git-svn-id: https://matlabulchifai.googlecode.com/svn/trunk@118 126f7d6c-fee3-11dd-88c7-29a3b14d5316