Added ability to view and delete user accounts from the admin section.  It currently blocks deletion of admin and tester accounts.  We will have to remove block on tester later.
Logout now invalidates the rails session and then flex forces a reload of the flash app.  This was the only way to get a new auth token to use the app again and it also prevents a users from hitting the back button and seeing sensitive data.

Updated lang array

git-svn-id: https://matlabulchifai.googlecode.com/svn/trunk@190 126f7d6c-fee3-11dd-88c7-29a3b14d5316