Fixed autocomplete "automatic selection" bug. Fixed potential bug when user deletes a record. Usually it take a while to delete a patient when a user clicks delete. The basic fix was to disable the screen until the service call came back. Also change color of menu bar and changed font size to 14. Changed overall font size to 12 for the whole app.
there still an issue for creating visit_types, allergies, etc. and it has to do with the ID maps. Tommy and I will continue to look into this.

We need to make sure we are consistent in the way we time things, i.e. how to deal with pop us and delayed service calls, and buttons with delayed delete service calls, etc. Now is a good time to look for bugs and inconsistencies with the UI like the font sizes being different and the column sizes that need a tad bit of tweaking.

if you click on the menu bar just outside the clickable area, a "you do not have permission" alert comes up.

remove debug panel and lines???

git-svn-id: https://matlabulchifai.googlecode.com/svn/trunk@202 126f7d6c-fee3-11dd-88c7-29a3b14d5316